namenode_rpc/rpc_test.go
package hadoop_rpc
import (
	"testing"
	"fmt"
	"bytes"
)

func TestNewHeaderPacket(t *testing.T) {
	h := NewHeaderPacket()

	//assert conditions on the different
	//defaults
	if h.version != 7 {
		t.FailNow();
		fmt.Println("h.version: ", h.version)
	}
	if string(h.header) != "hrpc" {
		t.FailNow();
		fmt.Println("h.header: ", string(h.header))
	}
	if h.auth_method != 80 {
		t.FailNow();
		fmt.Println("h.auth_method: ", string(h.auth_method))
	}
	if h.serialization_type != 0 {
		t.FailNow();
		fmt.Println("h.serialization_type: ", string(h.serialization_type))
	}
	fmt.Println("pass")
}

//Unit test for Headers()
//Needs some testing itself with netcat, not
//sure if it is a valid test yet
func TestHeaders(t *testing.T) {
  header := NewHeaderPacket()

  byte_packet := header.Headers()
  if byte_packet == nil {
  	t.FailNow();
  }

  //first four bytes are "hrpc"
  //fifth byte: version
  //sixth byte: auth method
  //seventh byte: serialization method
  expected := []byte{104, 114, 112, 99, 7, 80, 0}

  if bytes.Compare(byte_packet, expected) != 0 {
  	t.FailNow();
  	fmt.Println("byte_packet: ", byte_packet, "expected: ", expected)
  }

  fmt.Println("byte packet: ", byte_packet)
}
